requirements.txt
# Unit Test Automation Dependencies
groq>=0.37.0
datasets>=2.16.0
python-dotenv==1.0.0
numpy>=1.26.0
pytest==7.4.3
huggingface-hub>=0.19.4

